# Locate performance issue on MSVC