`pow` function has performance issue on MSVC, which is about 5x slower than the result of clang-cl.

MSVC:
```
f32_pow            41.2 ns         41.4 ns     16592593
```

clang-cl:
```
f32_pow            8.56 ns         8.58 ns     74666667
```

Besides, `fixed64` are generally slower on MSVC.